+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContentList/ComboboxContentList.tsx
@@ -3,6 +3,7 @@ import classnames from "classnames";
 import { Text } from "@jobber/components/Text";
 import { Glimmer } from "@jobber/components/Glimmer";
 import { Spinner } from "@jobber/components/Spinner";
+import { AnimatedPresence } from "@jobber/components/AnimatedPresence";
 import styles from "./ComboboxContentList.css";
 import {
   ComboboxListProps,
@@ -39,20 +40,20 @@ export function ComboboxContentList(props: ComboboxListProps): JSX.Element {
                 key={option.id}
                 id={option.id}
                 label={option.label}
+                prefix={option.prefix}
               />
             );
           })}
 
-          {props.loading && optionsExist && (
-            <div
-              className={classnames([
-                styles.loadingContainer,
-                styles.hasOptions,
-              ])}
-            >
-              <Spinner size="small" />
-            </div>
-          )}
+          <div
+            className={classnames([styles.loadingContainer, styles.hasOptions])}
+          >
+            <AnimatedPresence transition="fromBottom">
+              {Boolean(props.loading && optionsExist) && (
+                <Spinner size="small" />
+              )}
+            </AnimatedPresence>
+          </div>
 
           {props.onLoadMore && (
             <ComboboxLoadMore onLoadMore={props.onLoadMore} />