+++ packages/components/src/Grid/Grid.tsx
@@ -39,9 +39,12 @@ export function Grid({
 
   return (
     <div data-testid={GRID_TEST_ID} className={classnames}>
-      {React.Children.map(children, ({ props }: { props: GridCellProps }) => {
-        return <InternalGridCell {...props} />;
-      })}
+      {React.Children.map(
+        children,
+        ({ props }: { readonly props: GridCellProps }) => {
+          return <InternalGridCell {...props} />;
+        },
+      )}
     </div>
   );
 }