+++ packages/components/src/StatusIndicator/index.ts
@@ -1,0 +1,2 @@
+export { StatusIndicator } from "./StatusIndicator";
+export type { StatusIndicatorType } from "./StatusIndicator.type";