+++ packages/design/src/getMobileShadows.js
@@ -63,6 +63,7 @@ function getShadowStyles(baseTokens) {
     },
   };
   const iOSShadows = sharedShadows;
+
   return { androidShadows, iOSShadows };
 }