+++ packages/hooks/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.1](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@2.9.0...@jobber/hooks@2.9.1) (2024-03-08)
+
+
+### Bug Fixes
+
+* **eslint:** Cleanup lint warnings ([#1796](https://github.com/GetJobber/atlantis/issues/1796)) ([3e0c770](https://github.com/GetJobber/atlantis/commit/3e0c770eca0964636f79831016dbbe25926ae951))
+
+
+
+
+
+# [2.9.0](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@2.8.3...@jobber/hooks@2.9.0) (2024-03-01)
+
+
+### Features
+
+* **components:** Added accessible name for combobox trigger ([#1786](https://github.com/GetJobber/atlantis/issues/1786)) ([3280713](https://github.com/GetJobber/atlantis/commit/32807139ee436f45839edd261620e359d156cc80))
+
+
+
+
+
+## [2.8.3](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@2.8.2...@jobber/hooks@2.8.3) (2024-02-22)
+
+**Note:** Version bump only for package @jobber/hooks
+
+
+
+
+
 ## [2.8.2](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@2.8.1...@jobber/hooks@2.8.2) (2024-02-05)